// supervision-client/src/app/shared/models/invoice.ts
export class Invoice {
    id: number;
    invoiceDate: Date;
    invoiceNumber: number;
    paymentDate: Date;
    paymentStatus: string;
    title: string;
    title1: string;
    title2: string;
    amount: number;
    amount1: number;
    amount2: number;
    vatAmount: number;
    subscriptionStartDate: Date;
    subscriptionEndDate: Date;
    invoiceStatus: number;
}